Mark all printf-like functions with __attribute__((format(printf, ...)) attribute in compielrs that support this feature and fix errors that are highlighted by the compiler

+++ code/game/g_session.c
@@ -52,7 +52,7 @@ static void G_WriteClientSessionData(gclient_t* client)
            client->sess.losses,
            client->sess.teamLeader);
 
-    var = va("session%i", client - level.clients);
+    var = va("session%i", (int)(client - level.clients));
 
     trap_Cvar_Set(var, s);
 }
@@ -73,7 +73,7 @@ void G_ReadSessionData(gclient_t* client)
     int spectatorState;
     int sessionTeam;
 
-    var = va("session%i", client - level.clients);
+    var = va("session%i", (int)(client - level.clients));
     trap_Cvar_VariableStringBuffer(var, s, sizeof(s));
 
     sscanf(s, "%i %i %i %i %i %i %i",
