+++ code/server/sv_client.c
@@ -536,7 +536,7 @@ Abort a download if in progress
 static void SV_StopDownload_f(client_t* cl)
 {
     if (*cl->downloadName)
-        Com_DPrintf("clientDownload: %d : file \"%s\" aborted\n", cl - svs.clients, cl->downloadName);
+        Com_DPrintf("clientDownload: %d : file \"%s\" aborted\n", (int)(cl - svs.clients), cl->downloadName);
 
     SV_CloseDownload(cl);
 }
@@ -568,11 +568,11 @@ static void SV_NextDownload_f(client_t* cl)
     int block = atoi(Cmd_Argv(1));
 
     if (block == cl->downloadClientBlock) {
-        Com_DPrintf("clientDownload: %d : client acknowledge of block %d\n", cl - svs.clients, block);
+        Com_DPrintf("clientDownload: %d : client acknowledge of block %d\n", (int)(cl - svs.clients), block);
 
         // Find out if we are done.  A zero-length block indicates EOF
         if (cl->downloadBlockSize[cl->downloadClientBlock % MAX_DOWNLOAD_WINDOW] == 0) {
-            Com_Printf("clientDownload: %d : file \"%s\" completed\n", cl - svs.clients, cl->downloadName);
+            Com_Printf("clientDownload: %d : file \"%s\" completed\n", (int)(cl - svs.clients), cl->downloadName);
             SV_CloseDownload(cl);
             return;
         }
@@ -620,7 +620,7 @@ void SV_WriteDownloadToClient(client_t* cl, msg_t* msg)
     if (!cl->download) {
         // We open the file here
 
-        Com_Printf("clientDownload: %d : begining \"%s\"\n", cl - svs.clients, cl->downloadName);
+        Com_Printf("clientDownload: %d : begining \"%s\"\n", (int)(cl - svs.clients), cl->downloadName);
 
         missionPack = FS_idPak(cl->downloadName, "missionpack");
         idPack = missionPack || FS_idPak(cl->downloadName, "baseq3");
@@ -628,7 +628,7 @@ void SV_WriteDownloadToClient(client_t* cl, msg_t* msg)
         if (!sv_allowDownload->integer || idPack || (cl->downloadSize = FS_SV_FOpenFileRead(cl->downloadName, &cl->download)) <= 0) {
             // cannot auto-download file
             if (idPack) {
-                Com_Printf("clientDownload: %d : \"%s\" cannot download id pk3 files\n", cl - svs.clients, cl->downloadName);
+                Com_Printf("clientDownload: %d : \"%s\" cannot download id pk3 files\n", (int)(cl - svs.clients), cl->downloadName);
                 if (missionPack) {
                     Com_sprintf(errorMessage, sizeof(errorMessage), "Cannot autodownload Team Arena file \"%s\"\n"
                                                                     "The Team Arena mission pack can be found in your local game store.",
@@ -637,7 +637,7 @@ void SV_WriteDownloadToClient(client_t* cl, msg_t* msg)
                     Com_sprintf(errorMessage, sizeof(errorMessage), "Cannot autodownload id pk3 file \"%s\"", cl->downloadName);
                 }
             } else if (!sv_allowDownload->integer) {
-                Com_Printf("clientDownload: %d : \"%s\" download disabled", cl - svs.clients, cl->downloadName);
+                Com_Printf("clientDownload: %d : \"%s\" download disabled", (int)(cl - svs.clients), cl->downloadName);
                 if (sv_pure->integer) {
                     Com_sprintf(errorMessage, sizeof(errorMessage), "Could not download \"%s\" because autodownloading is disabled on the server.\n\n"
                                                                     "You will need to get this file elsewhere before you "
@@ -653,7 +653,7 @@ void SV_WriteDownloadToClient(client_t* cl, msg_t* msg)
             } else {
                 // NOTE TTimo this is NOT supposed to happen unless bug in our filesystem scheme?
                 //   if the pk3 is referenced, it must have been found somewhere in the filesystem
-                Com_Printf("clientDownload: %d : \"%s\" file not found on server\n", cl - svs.clients, cl->downloadName);
+                Com_Printf("clientDownload: %d : \"%s\" file not found on server\n", (int)(cl - svs.clients), cl->downloadName);
                 Com_sprintf(errorMessage, sizeof(errorMessage), "File \"%s\" not found on server for autodownloading.\n", cl->downloadName);
             }
             MSG_WriteByte(msg, svc_download);
@@ -762,7 +762,7 @@ void SV_WriteDownloadToClient(client_t* cl, msg_t* msg)
             MSG_WriteData(msg, cl->downloadBlocks[curindex], cl->downloadBlockSize[curindex]);
         }
 
-        Com_DPrintf("clientDownload: %d : writing block %d\n", cl - svs.clients, cl->downloadXmitBlock);
+        Com_DPrintf("clientDownload: %d : writing block %d\n", (int)(cl - svs.clients), cl->downloadXmitBlock);
 
         // Move on to the next block
         // It will get sent with next snap shot.  The rate will keep us in line.
@@ -1346,6 +1346,6 @@ void SV_ExecuteClientMessage(client_t* cl, msg_t* msg)
     } else if (c == clc_moveNoDelta) {
         SV_UserMove(cl, msg, false);
     } else if (c != clc_EOF) {
-        Com_Printf("WARNING: bad command byte for client %i\n", cl - svs.clients);
+        Com_Printf("WARNING: bad command byte for client %i\n", (int)(cl - svs.clients));
     }
 }
